Move some logic from the abstract class to the PossibleCallMatcher
